% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_telegram.R
\docType{class}
\name{client_telegram-class}
\alias{client_telegram-class}
\title{An S4 class representing Telegram client.}
\description{
Class extending a \code{\link{notifieR_client-class}} class for Telegram
client. In addition to any fields in the \code{\link{notifieR_client-class}}
class this one contains `telegram_token` which is needed to send a message
via Telegram API.
}
\section{Slots}{

\describe{
\item{\code{telegram_token}}{A telegram bot API token.}
}}

