% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_telegram.R
\name{client_telegram}
\alias{client_telegram}
\title{Telegram client's constructor.}
\usage{
client_telegram(telegram_token)
}
\arguments{
\item{telegram_token}{<ToDo>}
}
\description{
This function takes `telegram_token` and returns an object of class
\code{\link{client_telegram-class}}.
}
