% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifieR_client.R
\docType{class}
\name{notifieR_client-class}
\alias{notifieR_client-class}
\title{A main S4 class of notifieR package.}
\description{
<ToDo: description>
}
\section{Slots}{

\describe{
\item{\code{type}}{A type (name) of service whose class is wrapping \code{notifieR_client-class}.}
}}

