% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifieR_client.R
\name{notifieR_client}
\alias{notifieR_client}
\alias{is.notifieR_client}
\alias{send_message.notifieR_client}
\alias{print.notifieR_client}
\title{notifieR client}
\usage{
notifieR_client(service)

is.notifieR_client(x)

\method{send_message}{notifieR_client}(client, message, destination, ...)

\method{print}{notifieR_client}(x, ...)
}
\arguments{
\item{service}{a name of a client's serivce.}
}
\description{
Function creates an object of \code{notifieR_client} class with
given \code{service} field, typically a name of selected service. This object
is an underlying structure from which every other client is created.

\code{is.notifieR_client} checks if a provided object is of 
the \code{notifieR_client} class and whether it has all the fields
a \code{notifieR_client} should have.

\link{send_message} method for \code{notifieR_client} only
serves as a placeholder. For details on how to create a custom client
see \link{custom_client}.
}
\seealso{
\link{available_clients} \link{custom_client}
}
