% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{notify}
\alias{notify}
\title{Create a desktop notification}
\usage{
notify(msg)
}
\arguments{
\item{msg}{Message to show. It may contain newline characters.}
}
\description{
Create a desktop notification
}
\examples{
\dontrun{
notify("Hello world!")
}
}

