% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifieR_client.R
\docType{methods}
\name{print,notifieR_client-method}
\alias{print,notifieR_client-method}
\title{Print function for notifieR client.}
\usage{
\S4method{print}{notifieR_client}(x, ...)
}
\arguments{
\item{x}{<ToDo>}

\item{...}{<ToDo>}
}
\description{
<ToDo: description>
}
