% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_message.R
\name{send_message}
\alias{send_message}
\title{A main method for sending messages.}
\usage{
send_message(client, message, destination, ...)
}
\arguments{
\item{client}{<ToDo>}

\item{message}{<ToDo>}

\item{destination}{<ToDo>}

\item{...}{<ToDo>}
}
\description{
<ToDo>
}
