% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_message.R, R/client_telegram.R
\docType{methods}
\name{send_message}
\alias{send_message}
\alias{send_message,client_telegram-method}
\title{A main method for sending messages.}
\usage{
send_message(client, message, destination, ...)

\S4method{send_message}{client_telegram}(client, message, destination, ...)
}
\arguments{
\item{client}{<ToDo>}

\item{message}{<ToDo>}

\item{destination}{<ToDo>}

\item{...}{<ToDo>}
}
\description{
<ToDo>

<ToDo: desc>
}
